fixed unnecissary downscaling of pointLaneWidth in drawNicePoints
niceDots now limits the the lower bound of a plot to 0 if logScale is active and an errorBar dips below zero. Note that it will still throw an error if any data points are less than zero with logScaling.

removed print flags for debuging in niceBar.

Changed the way dRange is calculated in niceBar