Further tweaks to theme settings, especially pointLaneWidth
fixed niceBar to set width at 1/4 in the case of a single factor level.

fixed niceBar handling of legend labels when using stack.

Allowed for color fill options to be used if point options absent of only one is listed in niceDots.

As with niceBar, niceDots also sets width to 1/4 value in single factor settings. Previously plotted nothing.

Changes in np_options processing to allow for better handling of cases where stack==TRUE