Defined nicePlots output as a new class npData.
npData now has S3 variants of NicePlots functions to allow the graph to be rerendered with same of different set of options.

New S3 extention of plot to support S3 data with plotType used to control which plot is called.

niceDensity has major changes including support for theta an phi rotations of the surface plotting options. The useRgl option is now functional and will use rgl to plot surfaces if set to true and the library is available.