Added functional xypos data to npData options output for niceBox, niceDots and niceVio.
fixed a bug where the filtered data was stored in th npData object instead of the original data. This is important as by retaining the original size allows for associated factors and data to be added later from the same data set.

fixed a bug where beeswarm plots were not returning only the last loops worth of data.

the xypos data is designed to be used for interactive point exploration with shiny,